pyproject.toml
[project]
name = "local-llm-p-grade"
version = "0.1.0"
description = "Secure Local AI Framework with RBAC, PII masking, and monitoring."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "rich>=13.7.1",
    "textual>=0.58.0",
    "pydantic>=2.7.1",
    "pypdf>=4.2.0",
    "pandas>=2.2.2",
    "loguru>=0.7.2",
    "sqlcipher3>=0.5.2",
    "passlib[argon2]>=1.7.4",
    "argon2-cffi>=23.1.0",
    "presidio-analyzer>=2.2.351",
    "presidio-anonymizer>=2.2.351",
    "spacy>=3.7.4",
    "watchdog>=4.0.0",
    "jinja2>=3.1.4",
    "watchdog>=4.0.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "black>=24.4.2",
    "isort>=5.13.2",
    "mypy>=1.10.0",
]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
